Put drums logic in a separate .js file.
