+++ views/layouts/main.handlebars
@@ -74,6 +74,7 @@
 <script src='https://cdnjs.cloudflare.com/ajax/libs/jquery/3.5.1/jquery.min.js'
     integrity='sha512-bLT0Qm9VnAYZDflyKcBaQ2gg0hSYNQrJ8RilYldYQ1FxQYoCLtUjuuRuZo+fjqhx/qtq/1itJ0C2ejDxltZVFg=='
     crossorigin='anonymous'></script>
+<script type='text/javascript' src='../../assets/js/drums.js'></script>
 <script type='text/javascript' src='../../assets/js/signup.js'></script>
 <script type='text/javascript' src='../../assets/js/login.js'></script>
 <script type='text/javascript' src='../../assets/js/script.js'></script>
